feat: add useMultipleFileUploadHandler hook
